requirements.txt
python-telegram-bot[webhooks]==21.6
httpx>=0.27.0
python-dotenv==1.0.1
aiosqlite>=0.20.0
duckduckgo_search>=7.0.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
